requirements.txt
boto3==1.4.0
botocore==1.4.71
Django==1.10.3
django-cors-headers==1.3.1
django-polymorphic==1.0.2
djangorestframework==3.5.3
djangorestframework-jwt==1.8.0
docutils==0.12
jmespath==0.9.0
psycopg2==2.6.2
PyJWT==1.4.2
python-dateutil==2.6.0
s3transfer==0.1.2
six==1.10.0
